+++ R/cache.R
@@ -187,17 +187,35 @@ renv_cache_synchronize_impl <- function(cache, record, linkable, path) {
     renv_file_copy(path, cache)
   }
 
-  # change the cache owner if set
   if (renv_platform_unix()) {
+
+    # change the cache owner if set
     user <- Sys.getenv("RENV_CACHE_USER", unset = NA)
     if (!is.na(user)) {
       renv_system_exec(
         command = "chown",
         args    = c("-R", renv_shell_quote(user), renv_shell_path(cache)),
-        action  = "chowning cache folder",
+        action  = "chowning cached package",
         quiet   = TRUE
       )
     }
+
+    # change file modes after copy if set
+    mode <- Sys.getenv("RENV_CACHE_MODE", unset = NA)
+    if (!is.na(mode)) {
+      renv_system_exec(
+        command = "chmod",
+        args    = c("-R", renv_shell_quote(mode), renv_shell_path(cache)),
+        action  = "chmoding cached package",
+        quiet   = TRUE
+      )
+    }
+
+    # finally, allow for an arbitrary callback if set
+    callback <- getOption("renv.cache.callback")
+    if (is.function(callback))
+      callback(cache)
+
   }
 
   after <- Sys.time()